Add multicolumn index support.
Indexes used to be generated by looking at the column attributes for an
`Indexed` column, which limited the design to a single selector.  The new
method stores indexes in the table attributes which are then used by
`compileCreateIndex`.

API

The api for indexes has changed from accepting a Selector to a Group.
Thus code that used to appear `[col :- index]` might now need to be
`[Single col :- index]`.  With the advantage that we can now say
`[(Single col2 :+ Single col1) :- index]`.

The version of selda has been bumped in response to this API change.

Uniqueness

Index names are constructed by from the column names, which are
recovered by looking up each column info based on index. Any underscores
in the column name is escaped ("column_name" ~~> "column__name") and all
columns are joined using a single underscore to form the index name.